to_do_list/requirements.txt
django == 5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.1.1
psycopg2-binary==2.9.9
django-filter==24.2
drf-yasg==1.21.7
pytest==8.2.2
pytest-django==4.8.0

requests==2.32.4
datetime==5.5
pytz==2025.2